feat(products): refactor product model and implement GetProductsUsecase
